fix: correct Hytera EP keystream (40-bit MI) across all backends; guard SILENCE over-tagging (#4)
Hytera EP (ALG=0x02):
- compute_hytera_ks now uses kiv[i]=key5[i]^MI[i] over all 5 bytes with a 40-bit
  MI, matching DSD-FME hytera_enhanced_rc4_setup (was kiv[0]=key5[0] + 32-bit MI).
- Single-sourced in include/hytera_ks.h; CUDA (bruteforce.cu), OpenCL
  (dmrcrack.cl + bruteforce_cl.cpp) and CPU (bruteforce.c) all aligned. The
  OpenCL path had silently kept the old buggy formula.
- PayloadLine.mi widened to uint64_t; MOTOTRBO still uses the low 32 bits.
- Both converters (dsdfme_dsp_to_bin.py + dsp_convert_to_bin) parse the Hytera
  "DMR PI H- ... MI(40):" header, discriminating from MOTOTRBO by MI width.
- Validated by tests/test_hytera_ks.c (keystream vs DSD-FME spec).

SILENCE over-tagging (GitHub #4):
- load_payload_file ignores the KPA silence cache when >50% of frames are tagged
  SILENCE (untrustworthy converter heuristic), so the correct key is no longer
  pruned. Validated by tests/test_silence_guard.c.

GUI:
- Live-capture dialog redesigned for non-technical users: plain-language intro,
  frequency as the one prominent field, technical options behind "More options".

+++ include/payload_io.h
@@ -29,7 +29,7 @@ typedef struct {
     uint8_t silence_candidate; /* 1 = first voice burst of a new TX; voicing bit expected = 0 */
     uint8_t algid;
     uint8_t keyid;
-    uint32_t mi;
+    uint64_t mi;   /* up to 40-bit: MOTOTRBO uses low 32 bits, Hytera EP uses all 40 */
 } PayloadLine;
 
 typedef struct {
@@ -41,7 +41,7 @@ typedef struct {
     uint8_t has_global_keyid;
     uint8_t global_algid;
     uint8_t global_keyid;
-    uint32_t global_mi;
+    uint64_t global_mi;
     /* KPA silence frame index cache (built by load_payload_file) */
     uint16_t silence_indices[64];
     uint8_t  n_silence;          /* count, capped at 64 */
